Problem with Login
I have created the login functions and they are working through Postman but when I try to use them in a browser the website never responds.
What I am trying to do is use a session, a session is assigned if the user logs in with a valid username and password.